% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communities.R
\name{commuConsistency}
\alias{commuConsistency}
\title{Check Community Consistency in Population}
\usage{
commuConsistency(population)
}
\arguments{
\item{population}{population to check}
}
\description{
quick test that each community identifier (cid) in a given population
dataset adheres to certain consistency rules:
\enumerate{
\item Each cid, appears either once (dead partner) or twice in the population.
\item For each cid, all members of the population belonging to that cid
are part of the same community.
}
}
