% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communities.R
\name{countTraitsPerComu}
\alias{countTraitsPerComu}
\title{Count Traits Per Community}
\usage{
countTraitsPerComu(population, communities, traitsid)
}
\arguments{
\item{population}{A data frame containing information about individuals in a population, including their community and traits.}

\item{communities}{A data frame containing information about the communities in the population, including their size and number of adaptive traits.}

\item{traitsid}{A vector specifying the column indices of the traits in the population data frame.}
}
\value{
A matrix with the number of neutral and adaptive traits per community.
}
\description{
This function first aggregates the traits of individuals in the population data frame by community, using the community size as a factor. It then multiplies the number of adaptive traits in each community by the community size, and combines this with the number of neutral traits in each community. The result is a matrix with the number of neutral and adaptive traits per community.
}
