% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communities.R
\name{fissionCommunity}
\alias{fissionCommunity}
\title{Fission Community}
\usage{
fissionCommunity(comus, ol)
}
\arguments{
\item{comus}{A list representing the community,}

\item{ol}{The index of the community fissionning}
}
\value{
An updated list a new comunity
}
\description{
This function fission  community by identifying an unoccupied location .
}
