% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fullpathways}
\alias{fullpathways}
\title{All Paper Pathways}
\format{
A list with 4 elements describing the 45 pathways studied in the paper
\describe{
\item pre: which traits go through pre marital pathways
\item post: which traits go through post marital pathways
\item s are: the traits biased toward specific sex
\item tr: probability that the pathways is used
}
}
\usage{
fullpathways
}
\description{
a list describing all pathways used in the paper
Report ...
}
\keyword{datasets}
