% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socialLearning.R
\name{generatePathways}
\alias{generatePathways}
\title{Initialize Neutral Traits Pathways}
\usage{
generatePathways(z = 9, t_pre = c("v", "h", "o"), t_post = c("h", "o", "i"))
}
\arguments{
\item{z}{number of traitsj}

\item{t_pre}{pathway pre marital}

\item{t_post}{pathway post marital}
}
\description{
This function initializes the paramater matrix for the traits
}
\examples{
generatePathways(1, 5)
}
