% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socialLearning.R
\name{generateTraitsMatrix}
\alias{generateTraitsMatrix}
\title{Generate a Matrix of Traits for a Population}
\usage{
generateTraitsMatrix(N, z = 9, traitnames = "t", nastart = NULL)
}
\arguments{
\item{N}{An integer representing the number of individuals in the population.}

\item{z}{An optional integer representing the number of traits per individual, defaulting to 9.}

\item{traitnames}{A string representing the base name for the traits, which will be concatenated with their index.}

\item{nastart}{An optional parameter; if non-NULL, all trait values are initialized to NA.}
}
\value{
A matrix of size N x z, where each row represents an individual and each column a trait.
The traits are named according to 'traitnames' and their index.
}
\description{
This function creates a matrix representing various traits for a population.
Each row in the matrix represents an individual, and each column represents a different trait.
The traits are initialized either to a sample from c(0,1) or to NA, depending on whether 'nastart' is NULL or not.
}
\examples{
# Generate a 5 traits matrix for a population of 100 individuals
traits_matrix <- generateTraitsMatrix(100, z=5)

}
