% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socialLearning.R
\name{getAgeBand}
\alias{getAgeBand}
\title{Get Age Band}
\usage{
getAgeBand(age.peers, age.ref, threshold, pos = "h")
}
\arguments{
\item{age.peers}{A numeric vector representing peer ages to compare against \code{age.ref}.}

\item{age.ref}{A numeric value or vector representing the reference age(s) for comparison.}

\item{threshold}{A numeric value representing the threshold for age difference.}

\item{pos}{A character string specifying which peers are selected:
'h' horizontal copy ; the absolute difference in age is within the threshold,
'o' oblique,  \code{age.peers} is greater than or equal to \code{age.ref + threshold},
'b' below,  \code{age.peers} is less than or equal to \code{age.ref - threshold}.
Defaults to 'h'.}
}
\value{
A logical vector (if age is a unique value) or matrix (if ages is a vector) indicating whether each element in \code{age.peers} meets
the criteria specified by \code{pos} and \code{threshold} relative to \code{age.ref}.
}
\description{
This function compares ages in \code{age.peers} with a reference age \code{age.ref}
based on a specified \code{threshold}. The type of comparison is determined by the \code{pos} parameter.
}
