% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_tools.R
\name{getDensities}
\alias{getDensities}
\title{Calculate Density with Adjusted Bounds and Normalize}
\usage{
getDensities(traits, bw = 0.05)
}
\arguments{
\item{traits}{Numeric vector representing the traits to calculate the density of.}

\item{bw}{Bandwidth parameter for the density calculation; controls the smoothness.}

\item{na.rm}{Logical; if TRUE, missing values are removed before density calculation.}
}
\value{
A matrix with two columns. The first column contains the x-values, and the
second column contains the normalized y-values of the density.
}
\description{
This function calculates the density of given traits, adjusts the bounds of
the density calculation, and normalizes the density values.
}
\examples{
traits <- c(rnorm(100, 0.5, 0.1), rnorm(100, 0.7, 0.1))
densResult <- getDensities(traits)
plot(densResult[,1], densResult[,2], type = 'l')
}
