% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socialLearning.R
\name{getRatio}
\alias{getRatio}
\title{getRatio}
\usage{
getRatio(listtraits)
}
\arguments{
\item{listtraits}{a vector of traits 0 and 1}
}
\value{
ratio of 1 over the total number of traits
}
\description{
This function initializes the matrix for the traits of a population
}
