% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socialLearning.R
\name{initAdaptiveTraits}
\alias{initAdaptiveTraits}
\title{Initialize Adaptive Traits for Communities}
\usage{
initAdaptiveTraits(km, ki, aval = c(i = 0, m = 1), n = 3)
}
\arguments{
\item{km}{integer, the number of migrants communities}

\item{ki}{integer, the number of incumbents communities}

\item{aval}{named numeric vector with initial values for "m" (migrants) and "i" (incumbents) traits default c("i" = 0, "m" = 1).}

\item{n}{integer, the number of adaptive traits  default is 3.}
}
\value{
A matrix where each column represents an individual and each row represents a trait, with migrants and incumbents.
}
\description{
This function initializes the traits for each community, creating a matrix of traits for migrants and incumbents communities.
}
\examples{
initAdaptiveTraits(1, 5)
}
