% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communities.R
\name{initialiseCommunities}
\alias{initialiseCommunities}
\title{Initialise Communities with Traits and Coordinates}
\usage{
initialiseCommunities(
  coordinates = NULL,
  initcoor = "random",
  G = NULL,
  ks = NULL,
  traits = NULL,
  km = NULL,
  ki = NULL,
  K = NULL,
  migrantscount = NULL,
  sizes = NULL,
  plot = F
)
}
\arguments{
\item{coordinates}{matrix, a pre-defined set of coordinates for the communities; if NULL, coordinates will be generated based on other criteria.}

\item{initcoor}{character, if "random", coordinates are generated randomly; otherwise, a specified setup is expected.}

\item{G}{integer, used when coordinates are to be generated randomly, indicating the grid size.}

\item{ks}{numeric, the proportion of incumbents; used when \code{K} is defined but \code{ki} and \code{km} are not.}

\item{traits}{matrix, a pre-defined set of traits for the community; if NULL, traits will be generated internally.}

\item{km}{integer, the number of migrants communities.}

\item{ki}{integer, the number of incumbents communities.}

\item{K}{integer, the total number of communities ; if NULL, it's derived from \code{ki} and `km.}
}
\value{
A list containing matrices or vectors for the coordinates, adaptive traits, and sizes of each the communities.
}
\description{
This function sets up communities by initializing their spatial coordinates, adaptive traits, and sizes.
}
\examples{
comus=initialiseCommunities(G = 100, ki = 5, km = 5)
plot(comus$coordinates,pch=21,bg=apply(comus$adaptivetraits,1,mean)+2,xlim=c(0,100),ylim=c(0,100))
}
