% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socialLearning.R
\name{initNeutralTraits}
\alias{initNeutralTraits}
\title{Initialize Neutral Traits}
\usage{
initNeutralTraits(N, z = 9, traitnames = "t", nastart = NULL)
}
\arguments{
\item{N}{population size}

\item{z}{number of traitsj}

\item{traitnames}{name to give to traits, default t1 .. tz}

\item{nastart}{should the traits start as NA}
}
\description{
This function initializes the matrix for the traits of a population
}
\examples{
initNeutralTraits(1, 5)
}
