% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socialLearning.R
\name{initNeutralTraitsPathways}
\alias{initNeutralTraitsPathways}
\title{Initialize Neutral Traits Pathways}
\usage{
initNeutralTraitsPathways(
  z = 9,
  t_pre = c("v", "h", "o"),
  t_post = c("h", "o", "i")
)
}
\description{
This function initializes the paramater matrix for the traits
}
\examples{
initNeutralTraitsPathways(1, 5)
}
