% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joyplot.R
\name{joyplot}
\alias{joyplot}
\title{Joyplot Function}
\usage{
joyplot(
  listdensities,
  pathwaysnames = pathwaysnames,
  main = "",
  angle = 25,
  yadj = 2,
  sep_ba = c(0, 0.4),
  alpha = 0.8,
  baselines = F,
  miniaxes = F
)
}
\arguments{
\item{listdensities}{A list containing density distributions.}

\item{pathwaysnames}{A character vector representing names of pathways.}

\item{main}{Main title for the plot.}

\item{angle}{Angle of the peaks within each distribution.}

\item{yadj}{Adjustment parameter for the vertical spacing of the distributions.}

\item{sep_ba}{Separation between baseline plots.}

\item{alpha}{Alpha level for color transparency.}

\item{baselines}{Logical; if TRUE, baselines will be drawn for each distribution.}

\item{miniaxes}{Logical; if TRUE, mini-axes are drawn.}
}
\value{
Produces a plot.
}
\description{
This function creates a joyplot (also known as a ridgeline plot) to visualize distributions or densities.
}
