% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computerate.R
\name{lambda}
\alias{lambda}
\title{Calculate Adjusted Birth Rate}
\usage{
lambda(base_rate, bonus_rate, advantageous_traits)
}
\arguments{
\item{base_rate}{Numeric, the base birth rate of the community.}

\item{bonus_rate}{Numeric, the rate at which each full farmer increases the birth rate.}

\item{advantageous_traits}{Numeric vector, each element representing the presence (1) or absence (0)
of an advantageous trait within the community.}
}
\value{
Numeric, the adjusted birth rate after accounting for advantageous traits.
}
\description{
This function calculates the adjusted birth rate for a community by taking into account the
base birth rate, a bonus rate, and the sum of advantageous traits within the community.
}
\examples{
lambda(2, 0.5, c(1, 1, 0))
lambda(1, 0.3, c(1, 1, 1))
}
