% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelVector.R
\name{modelVector}
\alias{modelVector}
\title{Model a Population Vector over Time}
\usage{
modelVector(
  N,
  F_Th = NULL,
  ki,
  km,
  K,
  m,
  b,
  r,
  deathage = c(0, 5, 18, 40, 65, 85),
  rho = 0.5,
  d,
  maturity,
  endrepro,
  a,
  tp,
  age.threshold = 20,
  population,
  comus,
  logging = "time",
  tstep,
  ma = 1,
  traitsid,
  getfinalpop = FALSE,
  out = c("popsize", "popsumary"),
  beta = 0,
  vidfile = NULL,
  warn = FALSE,
  testdebug = FALSE,
  remarriage = FALSE,
  popcapsize = NULL,
  fracfiss = 0.5
)
}
\arguments{
\item{N}{overall population size}

\item{F_Th}{Threshold for community fission, NULL by default. When a community's size exceeds this threshold, it splits.}

\item{ki, km, K, m, b, r, rho, d, maturity, endrepro, a, tp, age.threshold, population, comus, tstep, ma, traitsid}{Specific parameters used in the population model. Set marriage, communities etc...}

\item{deathage}{use to compute age depedant variable}

\item{logging}{Controls the level of logging. implemented so far: "time" ,"demo" ,"pairing" ,"migrantscount" ,"fission" ,"visu" ,"done"}

\item{getfinalpop}{Logical flag to determine if the final population should be returned.}

\item{out}{Character vector specifying the types of output to be generated. Possible values include "popsize" for population size and "popsumary" for a summary of the population.  implemented so far "popsize" ,"popsumary" ,"weddings" ,"traitsumary" ,"migrsum" ,"comufull" ,"finalpop" ,"finalmigrants" ,"finalcomus"}

\item{beta}{parameter to adjust impact of migrant pop size model, with a default value of 0}

\item{vidfile}{Optional file path for saving video output, NULL by default.}

\item{warn}{Logical flag to control the display of warnings.}

\item{testdebug}{Logical flag to control stop test for debugging}

\item{remarriage}{Logical flag to control if widower can remarry or not}

\item{popcapsize}{Stop the population if it reach a certain size}

\item{fracfiss}{fraction of individual leaving}
}
\value{
A list containing various elements depending on the 'out' parameter. Elements can include population size, population summary, final population, and others as specified in 'out'.
}
\description{
This function simulates the evolution of a population over a given number of time steps.
It models various aspects such as marriages, migrations, community changes, and more.
}
\examples{
# Example usage of the modelVector function
#result <- modelVector(N = 100, ki = 1, km = 1, K = 5, m = 0.5, b = 0.1, r = 0.05, rho = 0.5, d = 0.01, 
#                     maturity = 18, endrepro = 50, a = 1, tp = tp_param, age.threshold = 20, 
#                     population = initial_population, comus = initial_comus, tstep = 10, 
#                     ma = 1, traitsid = 1:5, getfinalpop = TRUE, out = c("popsize", "popsumary"))
#mortality=c(0.15,0.01,0.01,0.02,0.05,1)'
}
