% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mortality}
\alias{mortality}
\title{mortality rate by ages}
\format{
\subsection{\code{mortality}}{
}
}
\usage{
mortality
}
\description{
a vector describing mortality rate by age
Report ...
}
\keyword{datasets}
