% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduction.R
\name{newpop}
\alias{newpop}
\title{Create a population matrix with specified attributes}
\usage{
newpop(n, age = NULL, minid = 0, cid = NULL, fid = NULL, community = NULL)
}
\arguments{
\item{n}{Integer; the number of individuals in the population.}

\item{age}{Numeric vector or NULL; the ages of the individuals. If set to 'random', ages are assigned randomly between 0 and 85. If NULL, all ages are set to 0.}

\item{minid}{Integer; the starting ID for the population. Used to offset the IDs of individuals.}

\item{cid}{Numeric vector or NULL; the community IDs for the individuals. If NULL, all individuals are assigned to community -1.}

\item{fid}{Numeric vector or NULL; the family IDs for the individuals. If NULL, all individuals are assigned a family ID of -1.}

\item{community}{Numeric vector or NULL; the community numbers for the individuals. If NULL, all individuals are assigned to community 1.}
}
\value{
A matrix with each row representing an individual and columns for 'id', 'age', 'partner', 'community', 'canrepro', 'cid', 'fid', 'justmarried', and 'sex'.
}
\description{
This function generates a matrix representing a population with each row corresponding to an individual. Attributes include age, community, family IDs, and more. The function allows customization of the population size, age distribution, community, and family structures.
}
\examples{
newpop(10, age = "random", minid = 100)
newpop(5, age = c(25, 30, 35, 40, 45), community = c(1, 2, 2, 3, 3))
}
