% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paperpopulation}
\alias{paperpopulation}
\title{Population used for the paper}
\format{
\subsection{\code{paperpopulation}}{

A fake populaiton with valid structure (couple, families and ages)
500 row (individuals)  and 13 columns
\describe{
\item{id}{unique id for every individual}
\item{cid}{unique id for each couple}
\item{fid}{unique id for each family}
\item{age}{age of individual}
\item{sex}{Sex of individual}
\item{partner}{id of partner}
\item{community}{Community of each individual}
\item{t1 ... t4}{Neutral traits fo each individual}
}
}
}
\usage{
paperpopulation
}
\description{
This population has 10 communities ; 500 individual and mean of exactly 50 people per community. This happened by chance. so better keep it
Report ...
}
\keyword{datasets}
