% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communities.R
\name{plot.comu}
\alias{plot.comu}
\title{Plot Community}
\usage{
\method{plot}{comu}(comus, color_gradient = NULL, vidfile = NULL)
}
\arguments{
\item{comus}{A data frame containing information about the community including coordinates, adaptive traits, and size of each species.}

\item{color_gradient}{A vector of colors used to represent the adaptive traits of each species. If not provided, a default color gradient from deep green to golden yellow will be used.}

\item{vidfile}{A string specifying the name of the video file to save the plot as. If not provided, the plot will not be saved as a video.}
}
\value{
A plot of the community with adaptive traits and size of each species represented by different colors and sizes.
}
\description{
This function plots a community with adaptive traits and size of each communities represented by different colors and sizes respectively.
}
