% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population}
\alias{population}
\title{generic population}
\format{
\subsection{\code{population}}{

A data frame with 7,240 rows and 60 columns:
}
}
\usage{
population
}
\description{
A generic exemple
Report ...
}
\keyword{datasets}
