% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communities.R
\name{random2Dgrid}
\alias{random2Dgrid}
\title{Generate Random 2D Grid Coordinates}
\usage{
random2Dgrid(K, Gx, Gy = NULL)
}
\arguments{
\item{K}{integer, the number of communities to locate}

\item{Gx}{integer, the size of the grid in the x-dimension.}

\item{Gy}{integer, the size of the grid in the y-dimension. If NULL, it assumes a square grid and uses Gx for the y-dimension.}
}
\value{
A 2-column matrix where each row represents the x and y coordinates of a community on the grid.
}
\description{
This function generates a set of random 2D coordinates, useful for initializing locations on a grid.
}
\examples{
com=random2Dgrid(4, 100,100)
plot(com,xlim=c(0,100),ylim=c(0,100),pch=20)
}
