% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communities.R
\name{reassignFamiliesToNewCommunityFIDs}
\alias{reassignFamiliesToNewCommunityFIDs}
\title{Reassign Families to New Community}
\usage{
reassignFamiliesToNewCommunityFIDs(comid, population, newsize, newid)
}
\arguments{
\item{comid}{The ID of the community to be processed.}

\item{population}{A dataframe containing population data,
including community IDs and family IDs.}

\item{newsize}{The cumulative size limit for the family selection.}

\item{newid}{The new community ID to assign.}
}
\value{
A modified population dataframe with updated community IDs for the selected families.
}
\description{
This function reassigns families from a specified community to a new community ID
once the cumulative size of these families reaches a specified limit. This version work \emph{if and only if}
\code{fid} identify children \emph{and} their parents
}
\examples{
# Assuming `population_data` is a dataframe with 'community' and 'fid' columns
# reassignFamiliesToNewCommunityFIDs(1, population_data, 800, 2)

}
