% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communities.R
\name{reassignFamiliesToNewCommunityNoFIDs}
\alias{reassignFamiliesToNewCommunityNoFIDs}
\title{Reassign Families to New Community}
\usage{
reassignFamiliesToNewCommunityNoFIDs(
  comid,
  population,
  newsize,
  newid,
  debug = FALSE
)
}
\arguments{
\item{comid}{The ID of the community to be processed.}

\item{population}{A dataframe containing population data,
including community IDs and family IDs.}

\item{newsize}{The cumulative size limit for the family selection.}

\item{newid}{The new community ID to assign.}
}
\value{
A modified population dataframe with updated community IDs for the selected families.
}
\description{
This function reassigns families from a specified community to a new community ID
once the cumulative size of these families reaches a specified limit. This version
recreate nuclear families using \code{cid} and \code{fid}, than split the families
}
\examples{
# Assuming `population_data` is a dataframe with 'community' and 'fid' columns
reassignFamiliesToNewCommunity(1, population_data, 800, 2)

}
