% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socialLearning.R
\name{sexbiascopy}
\alias{sexbiascopy}
\title{Sex-biased copying}
\usage{
sexbiascopy(ta, tb, sb)
}
\arguments{
\item{ta}{A vector of values selected with proba 1-sb}

\item{tb}{A vector of values selected with proba sb}

\item{sb}{A numeric vector of probabilities (between 0 and 1) indicating the bias  for each traits}
}
\value{
A vector of values selected from \code{ta} or \code{tb} based on the bias condition.
}
\description{
This function performs sex-biased copying based on bias proba
}
\examples{
sexbiascopy(c(t1=0,t2=0,t3=1), c(1,1,0), c(0.1,0.9,0.5))
barplot(apply(replicate(100,sexbiascopy(c(t1=0,t2=0,t3=1),c(1,1,0),c(.1,.9,.5))),1,table),legend=TRUE)
}
