% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduction.R
\name{validCouple}
\alias{validCouple}
\title{Check Validity of Couples Based on Age and Partner Status}
\usage{
validCouple(population, maturity, endrepro)
}
\arguments{
\item{population}{A data frame representing the population.
It must contain the columns 'age' and 'partner'.}

\item{maturity}{An integer representing the minimum age for reproductive maturity.}

\item{endrepro}{An integer representing the maximum age for reproduction.}
}
\value{
A logical vector indicating whether each individual in the population
is part of a valid couple.
}
\description{
This function determines if individuals in a population are part of a valid couple
based on age and partner status. A valid couple is defined as individuals whose
age is greater than or equal to the 'maturity' age and less than the 'endrepro'
age, and who have a partner (indicated by a 'partner' value greater than 0).
}
\examples{
# maturity and endrepro ages are defined
valid_couples <- validCouple(population, maturity=1, endrepro=65)

}
