% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socialLearning.R
\name{generateTraitsMatrix}
\alias{generateTraitsMatrix}
\title{Generate a Matrix of Traits for a Population}
\usage{
generateTraitsMatrix(N, z = 9, traitnames = "t", initval = c(0, 1))
}
\arguments{
\item{N}{An integer representing the number of individuals in the population.}

\item{z}{An optional integer representing the number of traits per individual, defaulting to 9.}

\item{traitnames}{A string representing the base name for the traits, which will be concatenated with their index.}

\item{initval}{An optional parameter; if a signle number all traits are initialised with this value, if vector of length > 1 values are sampled from the vector}
}
\value{
A matrix of size N x z, where each row represents an individual and each column a trait.
The traits are named according to 'traitnames' and their index.
}
\description{
This function creates a matrix representing various traits for a population.
Each row in the matrix represents an individual, and each column represents a different trait.
The value of traits are initialized using initval
}
\examples{
# Generate a 5 traits matrix for a population of 100 individuals
traits_matrix <- generateTraitsMatrix(100, z=5)

}
