% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socialLearning.R
\name{getRatio}
\alias{getRatio}
\title{ratio of traits equal to 1 other the total number of trait}
\usage{
getRatio(listtraits)
}
\arguments{
\item{listtraits}{a vector of traits 0 or 1}
}
\value{
the frequencie of 1 over 0 in listtraits
}
\description{
ratio of traits equal to 1 other the total number of trait
}
