% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mortality}
\alias{mortality}
\title{mortality rate by ages}
\format{
\subsection{\code{mortality}}{

A vector for probability of death for different age catgory
6 values
\describe{
6 categories of ages that need to be complated by deathage
}
}
}
\usage{
mortality
}
\description{
a vector describing mortality rate by age
Report ...
}
\keyword{datasets}
