% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population}
\alias{population}
\title{generic population}
\format{
\subsection{\code{population}}{

A fake populaiton with valid structure (couple, families and ages)
600 row (individuals)  and 13 columns
\describe{
\item{id}{unique id for every individual}
\item{cid}{unique id for each couple}
\item{fid}{unique id for each family}
\item{age}{age of individual}
\item{sex}{Sex of individual}
\item{partner}{id of partner}
\item{community}{Community of each individual}
\item{t1 ... t4}{Neutral traits fo each individual}
}
}
}
\usage{
population
}
\description{
A generic exemple
Report ...
}
\keyword{datasets}
