% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socialLearning.R
\name{social.learning}
\alias{social.learning}
\title{Social Learning}
\usage{
social.learning(x = NULL, when = "pre", pathways, threshold, traitsid = NULL)
}
\arguments{
\item{x}{matrix containing population of agents}

\item{when}{character defining whether the transmission is pre-marial ('pre') or post-marital ('post')}

\item{threshold}{integer defining age tresholds for distinguishing horrizontal and oblique transmission.}

\item{list}{containing the transmission pathways for each trait (generated using \code{initNeutralTraitsPathways()})}
}
\value{
An updated matrix of the population of agents
}
\description{
Core social learning routine
}
